MinigameSelector: support arbitrary realm count
I didn't want to extend this further. I fully expect Stonemaier to sell
promo realms for Rolling Realms as new games (or even expansions are
released). This modifies the `MinigameSelector` to properly render an
arbitrary number of realm buttons, though a different UI might need to
be considered if we add too many more rows. At worst, this will enable
support for "Pendulum" and "The Society", each of which are included in
the game's retail release but not included in this app at present.